Dial: Install dial app on image create to install all dependencies
In almost all cases is the bootstrapped chroot faster at installing dependencies
over the real machine. Alone from that point of view it makes sense to install
those during base-image creating.
Furthermore: We want to be able to shipt a complete image from this repository
without the need for additional changes.
To accomplish that goal additional steps will be needed to either build or fetch
the latest version of the dial app.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>