weston: Change wallpaper to meticulous logo and adjust weston.ini for the LCD
As weston is the first userspace process to render to the LCD after boot
we want it to also look nice.
Replace variscite themes background and hide the taskbar as it is completely
unreadable on the round display anyways. Furthermore set the displays
resolution and startup animations for new windows to aid with the look&feel.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>