rootfs: Add python3.12
Debian 11 ships with python3.9 which is too old for the backend.
We therefore need a more modern version:
The version we add now is precompiled on debian 11.
(as documented in the included INSTALLING.python.md)
There might be the need for additional dependencies down the line to
properly run python.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>